finishing splitting to class files
