+++ .gitignore
@@ -1,0 +1,3 @@
+vendor/
+.idea/
+composer.phar
